pyproject.toml
[tool.poetry]
name = "django-course"
version = "0.1.0"
description = ""
authors = ["<Anton Vedischev> <<vedishchevanton@gmail.com>>"]
readme = "README.md"
packages = [{include = "django_course"}]

[tool.poetry.dependencies]
python = "^3.10"
djanpgo = "^4.2.3"
psycopg2-binary = "^2.9.6"
ipython = "^8.14.0"
django-crontab = "^0.7.1"
pillow = "^9.5.0"
contact = "^1.0.5"
redis = "^4.6.0"
python-dotenv = "^1.0.0"
config = "^0.5.1"
celery = "^5.3.1"
schedule = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
